requirements.txt
google-adk
python-dotenv
google-adk[a2a]
python-dotenv